ml/training/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
mlflow>=2.7.0
optuna>=3.3.0
pyyaml>=6.0


